# contracts/control-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelAPI BoundaryLab control API",
    "version": "0.3.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "Implemented local single-operator control plane"
    }
  ],
  "security": [
    {
      "OperatorSession": []
    }
  ],
  "paths": {
    "/session": {
      "post": {
        "operationId": "createSession",
        "summary": "Local bootstrap login; Origin checked; request body not logged",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          }
        },
        "security": []
      },
      "get": {
        "operationId": "getSession",
        "summary": "Restore an authenticated local session",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoredSession"
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSession",
        "summary": "Invalidate session",
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ]
      }
    },
    "/targets": {
      "get": {
        "operationId": "listTargets",
        "summary": "Trusted configured targets",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Target"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "Runtime discovery and remediation capability disclosure",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/policy": {
      "get": {
        "operationId": "getPolicy",
        "summary": "Reviewed invoice permission policy and exact hash",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/spec": {
      "get": {
        "operationId": "getSpecification",
        "summary": "Bundled target contract and operation inventory",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/discovery/analyses": {
      "post": {
        "operationId": "createDiscoveryAnalysis",
        "summary": "Passively analyze OpenAPI plus optional HAR; 2 MB cap; no external refs",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryAnalysis"
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listDiscoveryAnalyses",
        "summary": "List persisted passive analyses",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscoveryAnalysis"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ]
      }
    },
    "/discovery/analyses/{analysis_id}/reviews": {
      "post": {
        "operationId": "createCandidateReview",
        "summary": "Append an approve/reject decision with rationale and immutable candidate hash",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReview"
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReviewRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listCandidateReviews",
        "summary": "Read the append-only candidate decision ledger",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateReview"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ]
      }
    },
...
    "/runs/{run_id}": {
      "get": {
        "operationId": "getRun",
        "summary": "Run state, scoped assessment and report",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Typed error; see documentation for 401/403/409/413/422/429/503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OperatorSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "boundarylab_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "request_id": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "operator": {
            "type": "string",
            "minLength": 1
          },
          "csrf_token": {
            "type": "string",
            "minLength": 1
          },
          "expires_in_seconds": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "operator",
          "csrf_token",
          "expires_in_seconds"
        ]
      },
...
      "Capabilities": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "discovery",
          "remediation"
        ]
      }
    }
  }
}